Create ItemWideVariableUpdateSetV1_0.xml
Example of variable created on a catalog item onLoad which can be accessed throughout the runtime of the item view (works for onChange AND onSubmit). 

Contains two variables manipulating a third one to showcase functionality. The item wide variable does not have to be displayed for it to be manupulated. 

Use cases:
* Query building using multiple variables and or condition.
* Query building using predefined strings.
* Reuse and survivability of variable values and or objects.
* Conditional mandatory checkers.
* and many many mooore!